requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
tenacity==8.2.3
reportlab==4.0.7